Minor fixes.
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk@410 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/Agg/AggPopupHint.cpp
@@ -77,7 +77,7 @@ void AggPopupHint::renewHintGeometry ()
 
     size_ = QSize( width, height );
 
-    if( buf_->width() < size_.width() )
+    if( (int)buf_->width() < size_.width() )
         buf_->init( size_.width(), size_.height() );
 }
